Split AbstractFFTFilterImg
Add AbstractFilterImg.  Move common functionality of
AbstractFFtFilterImg to AbstractFilterImg.  FFT filters should derive
from AbstractFFTFilterImg.  Non-fft filters can derive from
AbstractFilterImg.
